+++ src/main/java/org/truffleruby/core/kernel/KernelNodes.java
@@ -43,7 +43,7 @@ import org.truffleruby.core.binding.RubyBinding;
 import org.truffleruby.core.cast.BooleanCastNode;
 import org.truffleruby.core.cast.BooleanCastNodeGen;
 import org.truffleruby.core.cast.BooleanCastWithDefaultNode;
-import org.truffleruby.core.cast.DurationToMillisecondsNodeGen;
+import org.truffleruby.core.cast.DurationToNanoSecondsNodeGen;
 import org.truffleruby.core.cast.NameToJavaStringNode;
 import org.truffleruby.core.cast.ToIntNode;
 import org.truffleruby.core.cast.ToStrNode;
@@ -1608,13 +1608,13 @@ public abstract class KernelNodes {
 
         @CreateCast("duration")
         protected RubyBaseNodeWithExecute coerceDuration(RubyBaseNodeWithExecute duration) {
-            return DurationToMillisecondsNodeGen.create(false, duration);
+            return DurationToNanoSecondsNodeGen.create(false, duration);
         }
 
         @Specialization
-        protected long sleep(long durationInMillis,
+        protected long sleep(long durationInNanos,
                 @Cached BranchProfile errorProfile) {
-            if (durationInMillis < 0) {
+            if (durationInNanos < 0) {
                 errorProfile.enter();
                 throw new RaiseException(
                         getContext(),
@@ -1627,13 +1627,13 @@ public abstract class KernelNodes {
             // it should only be considered if we are inside the sleep when Thread#{run,wakeup} is called.
             thread.wakeUp.set(false);
 
-            return sleepFor(getContext(), thread, durationInMillis, this);
+            return sleepFor(getContext(), thread, durationInNanos, this);
         }
 
         @TruffleBoundary
-        public static long sleepFor(RubyContext context, RubyThread thread, long durationInMillis,
+        public static long sleepFor(RubyContext context, RubyThread thread, long durationInNanos,
                 Node currentNode) {
-            assert durationInMillis >= 0;
+            assert durationInNanos >= 0;
 
             // We want a monotonic clock to measure sleep duration
             final long startInNanos = System.nanoTime();
@@ -1641,13 +1641,12 @@ public abstract class KernelNodes {
             context.getThreadManager().runUntilResult(currentNode, () -> {
                 final long nowInNanos = System.nanoTime();
                 final long sleptInNanos = nowInNanos - startInNanos;
-                final long sleptInMillis = TimeUnit.NANOSECONDS.toMillis(sleptInNanos);
 
-                if (sleptInMillis >= durationInMillis || thread.wakeUp.getAndSet(false)) {
+                if (sleptInNanos >= durationInNanos || thread.wakeUp.getAndSet(false)) {
                     return BlockingAction.SUCCESS;
                 }
 
-                Thread.sleep(durationInMillis - sleptInMillis);
+                TimeUnit.NANOSECONDS.sleep(durationInNanos - sleptInNanos);
                 return BlockingAction.SUCCESS;
             });