+++ src/main/ruby/truffleruby/core/truffle/kernel_operations.rb
@@ -22,12 +22,12 @@ module Truffle
       enum.to_enum(method) { enum.send(size_method) }
     end
 
-    def self.convert_duration_to_milliseconds(duration)
+    def self.convert_duration_to_nanoseconds(duration)
       unless duration.respond_to?(:divmod)
         raise TypeError, "can't convert #{duration.class} into time interval"
       end
       a, b = duration.divmod(1)
-      ((a + b) * 1000)
+      ((a + b) * 1_000_000_000)
     end
 
     def self.define_hooked_variable(name, getter, setter, defined = proc { 'global-variable' })