+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Bug fixes:
 * Fix `File::Stat`'s `#executable?` and `#executable_real?` predicates that unconditionally returned `true` for a superuser (#2690, @andrykonchin).
 * The `strip` option `--keep-section=.llvmbc` is not supported on macOS (#2697, @eregon).
 * Disallow the marshaling of polyglot exceptions since we can't properly reconstruct them (@nirvdrum).
+* Fix `Kernel#sleep` and `Mutex#sleep` for durations smaller than 1 millisecond (#2716, @eregon).
 
 Compatibility: